uses unit_amount_decimal to send invoice item amount to Stripe, updated type-def

+++ src/graphql-server/type-defs.js
@@ -277,7 +277,7 @@ module.exports = gql`
   }
 
   input StripeInvoiceItemInput {
-    unit_amount: Int
+    unit_amount: Float
     quantity: Int
   }
